add basic support for the mouse
Basic support for the mouse is provided to be able to activate a frame
when clicking on it.

To activate the behavior, the mousefocuspolicy variable must be set:

set mousefocuspolicy click

Currently mousefocuspolicy only accept two values: none (no mouse
support at all) and click (activate a frame on click).